requirements-dev.txt
fake-useragent==1.5.1
requests==2.31.0
aiohttp==3.9.5
pytest-playwright==0.5.0
ruff==0.5.2
mypy==1.10.1
pre-commit==3.8.0
Appium-Python-Client==4.3.0
curl-cffi==0.7.4
zendriver==0.4.1
selenium==4.26.1
pytest_mock==3.14.0
pytest-asyncio==0.26.0
